Rename drbd.py to drbdmod.py to unified style.
